[*] Update Server Engine

+++ Server/Src/LockstepEngine/Src/LockstepEngine/ECS.Common/Interfaces/Services/INetworkService.cs
@@ -6,7 +6,7 @@ namespace Lockstep.Game {
     public interface INetworkService :IService {
         void SendGameEvent(byte[] data);
         void SendInput(Msg_PlayerInput msg);
-        void SendHashCodes(int startTick,List<long> hashCodes,int startIdx,int count);
+        void SendHashCodes(int startTick,List<int> hashCodes,int startIdx,int count);
         void SendMissFrameReq(int missFrameTick);
         void SendMissFrameRepAck(int missFrameTick);
     }
